requirements.txt
asgiref==3.6.0
async-generator==1.10
attrs==22.2.0
beautifulsoup4==4.11.2
branca==0.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
colorclass==2.2.2
contourpy==1.0.7
cycler==0.11.0
data==0.4
decorator==5.1.1
Django==4.1.7
django-filter==22.1
djangorestframework==3.14.0
docopt==0.6.2
folium==0.14.0
fonttools==4.38.0
funcsigs==1.0.2
gpxpy==1.5.0
h11==0.14.0
idna==3.4
Jinja2==3.1.2
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.7.0
numpy==1.24.2
outcome==1.2.0
packaging==23.0
Pillow==9.4.0
pip-update==0.4.1
pip-upgrader==1.4.15
polyline==2.0.0
psycopg2-binary==2.9.5
pyparsing==3.0.9
PySocks==1.7.1
pyTelegramBotAPI==4.10.0
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
requests==2.28.2
selenium==4.8.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
sqlparse==0.4.3
terminaltables==3.1.10
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.14
wsproto==1.2.0
